services/requirements.txt
# Core dependencies for Arealis Gateway services
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# AWS
boto3==1.34.0
botocore==1.34.0

# HTTP requests
requests==2.31.0
httpx==0.25.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2

# Logging
structlog==23.2.0

# JSON handling
orjson==3.9.10

# Decimal handling
decimal==1.70
